job_runner: Jobs now run from software.
job_runner.py: Now this file is able to launch jobs from software without
having to execute in a shell.  This commit blocks on I/O with the cluster
and returns a generator to each line produced by the output.

To Run the example
$ ./job_runner.py - This will configure your local environment for a job,
spin up a cluster pool if one doesn't exist, attach the job to an existing
pool and block till completion.